+++ packages/liveblocks-core/src/LiveList.ts
@@ -3,13 +3,9 @@ import { AbstractCrdt, OpSource } from "./AbstractCrdt";
 import { nn } from "./lib/assert";
 import { comparePosition, makePosition } from "./lib/position";
 import { LiveRegister } from "./LiveRegister";
-import type {
-  LiveListUpdateDelta,
-  LiveListUpdates,
-  LiveNode,
-  Lson,
-} from "./types";
+import type { LiveListUpdateDelta, LiveListUpdates } from "./types";
 import type { ToImmutable } from "./types/Immutable";
+import type { LiveNode, Lson } from "./types/Lson";
 import type { ParentToChildNodeMap } from "./types/NodeMap";
 import type { CreateChildOp, CreateListOp, CreateOp, Op } from "./types/Op";
 import { OpCode } from "./types/Op";
