+++ packages/liveblocks-core/src/__tests__/immutable.test.ts
@@ -7,16 +7,12 @@ import {
 } from "../immutable";
 import * as console from "../lib/fancy-console";
 import { LiveObject } from "../LiveObject";
-import type {
-  Json,
-  JsonObject,
-  LsonObject,
-  StorageUpdate,
-  ToJson,
-} from "../types";
+import type { StorageUpdate } from "../types";
 import type { BaseUserMeta } from "../types/BaseUserMeta";
 import type { ClientMsg } from "../types/ClientMsg";
 import { ClientMsgCode } from "../types/ClientMsg";
+import type { Json, JsonObject } from "../types/Json";
+import type { LsonObject, ToJson } from "../types/Lson";
 import type { IdTuple, SerializedCrdt } from "../types/SerializedCrdt";
 import { ServerMsgCode } from "../types/ServerMsg";
 import {
