+++ packages/liveblocks-core/src/types/index.ts
@@ -742,5 +742,3 @@ export enum WebsocketCloseCodes {
 }
 
 export type { Immutable, ToImmutable } from "./Immutable";
-export type { Json, JsonObject } from "./Json";
-export type { LiveNode, LiveStructure, Lson, LsonObject, ToJson } from "./Lson";
